+++ Plugin/src/main/java/ua/nanit/limbo/protocol/packets/play/PacketSpawnPosition.java
@@ -12,6 +12,7 @@ import ua.nanit.limbo.protocol.registry.Version;
 
 public class PacketSpawnPosition implements PacketOut {
 
+    private String dimensionName = "minecraft:overworld";
     private long x;
     private long y;
     private long z;
@@ -19,6 +20,11 @@ public class PacketSpawnPosition implements PacketOut {
     public PacketSpawnPosition() { }
 
     public PacketSpawnPosition(long x, long y, long z) {
+        this("minecraft:overworld", x, y, z);
+    }
+
+    public PacketSpawnPosition(String dimensionName, long x, long y, long z) {
+        this.dimensionName = dimensionName;
         this.x = x;
         this.y = y;
         this.z = z;
@@ -26,8 +32,14 @@ public class PacketSpawnPosition implements PacketOut {
 
     @Override
     public void encode(ByteMessage msg, Version version) {
+        if (version.moreOrEqual(Version.V1_21_9)) {
+            msg.writeString(dimensionName);
+        }
         msg.writeLong(encodePosition(x, y ,z));
         msg.writeFloat(0);
+        if (version.moreOrEqual(Version.V1_21_9)) {
+            msg.writeFloat(0);
+        }
     }
 
     @Override