+++ Plugin/src/main/java/xyz/xreatlabs/nexauth/common/log/LogFilter.java
@@ -35,8 +35,6 @@ public abstract class LogFilter {
     }
 
     protected boolean checkMessage(String message) {
-        // This sucks, but it's the only way to filter out the spam from the plugin
-        if (message.contains("Plugin listener xyz.xreatlabs.nexauth.bungeecord.BungeeCordListener took")) return false;
         if (!message.contains("issued server command: /") && !message.contains("executed command /") && !message.contains("executed command: /") && !message.contains("Duplicate key name"))
             return true;